Building BERT Model that predicts sentence affect
Building BERT Model that predicts whether the sentence has high or low affect on people's emotions i.e. high or low arousal. The dataset which was used is CrowdFlower and features were extracted from it with the help of NRC-VAD lexicon. As lower boundary for high affect is 0.8 and upper boundary for low affect is 0.442.